🧹 [Code Health] Extract duplicated rank override parsing logic into utility function
